fix: pulled session establishment out of RunCommand and into factory func
This makes more sense. It reduces the logic in the RunCommand method adhering to the single responsibility principle.

Note: We must ensure not to cause a memory leak. This is achieved by attaching the close methods of client and session to the sshClient struct.

All tests pass:
chrishern@PEARSONML-KAS9MD6 devcon % go test
PASS
ok      github.com/montybeatnik/devcon  2.737s